fixed warnings
git-svn-id: https://svn.opendtect.org/od/trunk@27480 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/uiEarthModel/uihorgeom2attr.cc
@@ -32,7 +32,7 @@ static const char* rcsID mUsedVar = "$Id$";
 	msfld_->attach( alignedBelow, att ); \
     }
 #define mGetZFac(valifms) \
-    const float zfac = msfld_ && msfld_->getBoolValue() ? valifms : 1
+    const float zfac = (float) (msfld_ && msfld_->getBoolValue() ? valifms : 1)
 
 
 uiHorGeom2Attr::uiHorGeom2Attr( uiParent* p, EM::Horizon3D& hor )
